// packages/destination-actions/src/destinations/marketo-static-lists/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Settings {
  /**
   * Your Marketo REST API Client ID.
   */
  client_id: string
  /**
   * Your Marketo REST API Client Secret.
   */
  client_secret: string
  /**
   * Your Marketo REST API Endpoint in this format: https://<your_account_id>.mktorest.com.
   */
  api_endpoint: string
  /**
   * Name of the folder in which to create static lists.
   */
  folder_name: string
}